refactor: 공통사용 되지 않는 컴포넌트 pages로 이동
