+++ src/pages/CreatePostPage/sections/CreatePostForm.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { css } from '@emotion/react';
-import ImageDropZone from '../../../components/common/ImageDropZone';
+import ImageDropZone from './ImageDropZone';
 import useInputs from '../../../hooks/useInputs';
 import { CreatePost, CreatePostImage } from '../../../types/post';
